key.cpp
Users of openssl implementation that do not have the NID_secp256k elliptical curve will need this patch to avoid the ,"CECKey::CECKey(): Assertion `pkey != __null' failed", at runtime. The main group effected would be redhat base systems such as centos,fedora,redhat and some opensuse environments.The reason the curve is not included in these system is related to patenting(they don't own the patent).If your receiving the error above or similar ,you would want to replace the existing key.cpp in the original source with this one.Run, 'make clean all', or remove key.o from the build directory prior to recompiling as this will force key.cpp to be rebuilt.